Refactor: Update `addtolist` screen and `RemoveFromList` component
- Modified `RemoveFromList.kt` to correct the package name and remove the `EmptyListsMessage` composable.
- Updated `WatchListItemUiState` in `addtolist.kt` to include `isLoading` and `itemCount` fields, and adjusted the mapping logic in the ViewModel accordingly.
- Removed an unused import for `RemoveFromListContent` from `addToList.kt`.
